create new Component NoSearch
