+++ src/components/main-page/main-page.jsx
@@ -1,9 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import {connect} from 'react-redux';
-
-import PlaceList from '../place-list/place-list.jsx';
-import Map from '../map/map.jsx';
 import CityList from '../city-list/city-list.jsx';
 
 import witchActiveItem from '../../hocs/witch-active-item/witch-active-item';
@@ -15,18 +12,17 @@ import {getCurrentOffers, getCitiesList} from '../../reducer/data/selectors.js';
 
 import {getCurrentCity} from '../../reducer/user/selectors.js';
 import Header from '../header/header.jsx';
-import SortingOption from '../sorting-option/sorting-option.jsx';
 
-import witchSortingOption from '../../hocs/witch-sorting-option/witch-sorting-option';
+import PlacesWrapped from '../../components/places-wrapped/places-wrapped.jsx';
+import NoResultsPage from '../../components/no-results/no-results-page.jsx';
 
-const PlaceListWrapped = witchActiveItem(PlaceList);
 const CityListWrapped = witchActiveItem(CityList);
-const SortingOptionWrapped = witchSortingOption(SortingOption);
+
 
 const MainPage = ({offers, cities, onChangeCity, currentCity}) => {
   return <div>
     <Header />
-    <main className="page__main page__main--index">
+    <main className={`page__main page__main--index ${!offers ? `page__main--index-empty` : ``}`}>
       <h1 className="visually-hidden">Cities</h1>
       <div className="cities tabs">
         <section className="locations container">
@@ -34,28 +30,9 @@ const MainPage = ({offers, cities, onChangeCity, currentCity}) => {
             onChangeCity={onChangeCity}
             currentCity={currentCity}
             cities={cities}/>
-        </section></div>
-      <div className="cities__places-wrapper">
-        <div className="cities__places-container container">
-          <section className="cities__places places">
-
-            <h2 className="visually-hidden">Places</h2>
-            <b className="places__found">{`${offers.length} places to stay in ${currentCity}`}</b>
-
-            <SortingOptionWrapped />
-
-            <PlaceListWrapped
-              offers={offers} />
-
-          </section>
-          <div className="cities__right-section">
-            <section className="cities__map map" >
-              <Map offers={offers}/>
-            </section>
-
-          </div>
-        </div>
+        </section>
       </div>
+      {!offers ? <NoResultsPage /> : <PlacesWrapped offers={offers} currentCity={currentCity}/> }
     </main>
   </div>;
 };
